refactor: reorganize error handling and update service trait
