+++ amico-sdk/src/ai/mod.rs
@@ -1,5 +1,6 @@
-pub mod model;
 pub mod embedding;
-pub mod service;
+pub mod errors;
+pub mod model;
 pub mod provider;
+pub mod service;
 pub mod tool;
